Don't initialize pitch before for loop Fix pipeline state's potential memory leak Add static cast to float

+++ ouzel/osx/OpenGLView.mm
@@ -137,7 +137,8 @@ using namespace ouzel;
 
     [_openGLContext update];
 
-    sharedEngine->getWindow()->setSize(Size2(newSize.width, newSize.height));
+    sharedEngine->getWindow()->setSize(Size2(static_cast<float>(newSize.width),
+                                             static_cast<float>(newSize.height)));
 
     CGLUnlockContext([_openGLContext CGLContextObj]);
 }
